+++ x-pack/plugins/observability_solution/slo/public/components/slo/error_rate_chart/use_lens_definition.ts
@@ -9,15 +9,12 @@ import { transparentize, useEuiTheme } from '@elastic/eui';
 import numeral from '@elastic/numeral';
 import { i18n } from '@kbn/i18n';
 import { TypedLensByValueInput } from '@kbn/lens-plugin/public';
-import {
-  ALL_VALUE,
-  SLOWithSummaryResponse,
-  timeslicesBudgetingMethodSchema,
-} from '@kbn/slo-schema';
+import { ALL_VALUE, SLOWithSummaryResponse } from '@kbn/slo-schema';
 import moment from 'moment';
 import { v4 as uuidv4 } from 'uuid';
 import { SLO_DESTINATION_INDEX_PATTERN } from '../../../../common/constants';
 import { SloTabId } from '../../../pages/slo_details/components/slo_details';
+import { getLensDefinitionInterval } from './utils';
 
 export interface TimeRange {
   from: Date;
@@ -48,16 +45,7 @@ export function useLensDefinition({
 }): TypedLensByValueInput['attributes'] {
   const { euiTheme } = useEuiTheme();
 
-  // For a timeslice SLO
-  // When the date range is lower than 24 hours, we force the interval to be the timeslice window
-  // Greater than 24 hours, the 'auto' interval will be fine
-  const dataTimeRangeDuration = moment(dataTimeRange.to).diff(moment(dataTimeRange.from), 'ms');
-  const interval =
-    timeslicesBudgetingMethodSchema.is(slo.budgetingMethod) &&
-    dataTimeRangeDuration <= 86400000 &&
-    slo.objective?.timesliceWindow
-      ? slo.objective.timesliceWindow
-      : 'auto';
+  const interval = getLensDefinitionInterval(dataTimeRange, slo);
 
   return {
     title: 'SLO Error Rate',
